change from int to float64 so that we can add less than 1 on the distance

+++ mapping_test.go
@@ -9,15 +9,15 @@ func TestGetCenterLocation(t *testing.T) {
 	testObjects := []struct {
 		Lat                    float64
 		Lon                    float64
-		Distance               int
-		LimitLength            int
+		Distance               float64
+		LimitLength            float64
 		TotalExpectedLocations int
 	}{
 		// this should be 100 locations
-		{Lat: -6.9875393, Lon: 108.4446289, Distance: 5, LimitLength: 50, TotalExpectedLocations: 100},
+		{Lat: -6.9875393, Lon: 108.4446289, Distance: 5.0, LimitLength: 50.0, TotalExpectedLocations: 100},
 
 		// this should be 25 locations
-		{Lat: -6.9875393, Lon: 108.4446289, Distance: 2, LimitLength: 10, TotalExpectedLocations: 25},
+		{Lat: -6.9875393, Lon: 108.4446289, Distance: 2.0, LimitLength: 10.0, TotalExpectedLocations: 25},
 	}
 
 	for _, test := range testObjects {
@@ -33,15 +33,15 @@ func TestGenerateLocation(t *testing.T) {
 	testObjects := []struct {
 		Lat                    float64
 		Lon                    float64
-		Distance               int
-		LimitLength            int
+		Distance               float64
+		LimitLength            float64
 		TotalExpectedLocations int
 	}{
 		// this should be 100 locations
-		{Lat: -6.9875393, Lon: 108.4446289, Distance: 5, LimitLength: 50, TotalExpectedLocations: 100},
+		{Lat: -6.9875393, Lon: 108.4446289, Distance: 5.0, LimitLength: 50.0, TotalExpectedLocations: 100},
 
 		// this should be 25 locations
-		{Lat: -6.9875393, Lon: 108.4446289, Distance: 2, LimitLength: 10, TotalExpectedLocations: 25},
+		{Lat: -6.9875393, Lon: 108.4446289, Distance: 2.0, LimitLength: 10.0, TotalExpectedLocations: 25},
 	}
 
 	for _, test := range testObjects {
@@ -58,13 +58,13 @@ func TestNewPoint(t *testing.T) {
 	testObjects := []struct {
 		Lat       float64
 		Lon       float64
-		Distance  int
+		Distance  float64
 		Direction string
 	}{
-		{Lat: -6.9875393, Lon: 108.4446289, Distance: 5, Direction: "west"},
-		{Lat: -6.9875393, Lon: 108.4446289, Distance: 5, Direction: "east"},
-		{Lat: -6.9875393, Lon: 108.4446289, Distance: 5, Direction: "north"},
-		{Lat: -6.9875393, Lon: 108.4446289, Distance: 5, Direction: "south"},
+		{Lat: -6.9875393, Lon: 108.4446289, Distance: 5.0, Direction: "west"},
+		{Lat: -6.9875393, Lon: 108.4446289, Distance: 5.0, Direction: "east"},
+		{Lat: -6.9875393, Lon: 108.4446289, Distance: 5.0, Direction: "north"},
+		{Lat: -6.9875393, Lon: 108.4446289, Distance: 5.0, Direction: "south"},
 	}
 
 	for _, test := range testObjects {
